Develop

+++ resources/views/cart/index.blade.php
@@ -51,11 +51,13 @@
                                     <div class="card-body d-flex justify-content-between">
                                         <div class="">
                                             <h5 class="card-title">{{ $item['name'] }}</h5>
-                                            <p class="card-text">R$ {{ $item['price'] }}</p>
+                                            <p class="card-text">
+                                                <span class="fw-bold">Qntd.:</span> {{ $item['quantity'] }} | <span class="fw-bold">R$</span> {{ $item['price'] }}
+                                            </p>
                                         </div>
 
                                         <div class="d-flex align-items-center">
-                                            <button class="btn btn-danger remove-from-cart" data-id="{{ $item['id'] }}">
+                                            <button class="btn btn-danger remove-from-cart" data-id="{{ $item['id'] }}" data-bs-toggle="modal" data-bs-target="#removeFromCart">
                                                 <i class="fa fa-trash" aria-hidden="true"></i>
                                             </button>
                                         </div>
@@ -100,6 +102,20 @@
         </div>
     </div>
 
+    <div class="modal fade" id="removeFromCart" tabindex="-1" aria-labelledby="removeFromCartLabel" aria-hidden="true">
+        <div class="modal-dialog">
+            <div class="modal-content">
+                <div class="modal-body d-flex justify-content-center">
+                    Deseja remover o produto do carrinho?
+                </div>
+                <div class="modal-footer justify-content-center">
+                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancelar</button>
+                    <button type="button" class="btn btn-danger" id="confirm-remove-from-cart">Remover</button>
+                </div>
+            </div>
+        </div>
+    </div>
+
     <footer class="bg-dark text-white py-3">
         <div class="container d-flex justify-content-center">
             Desenvolvido por Gabriel Ribeiro.
@@ -116,11 +132,21 @@
     <script>
         loadCart();
 
-        function removeProductFromCart() {
-            document.querySelectorAll(".remove-from-cart").forEach(button => {
-                button.addEventListener("click", function () {
-                    const productId = this.getAttribute('data-id');
+        let productId = null;
 
+        function removeProductFromCart() {            
+            document.querySelectorAll(".remove-from-cart").forEach(button => {
+                button.addEventListener("click", function() {
+                    productId = this.getAttribute("data-id");
+                });
+            });
+
+            requestToRemoveProductFromCart();
+        }
+
+        function requestToRemoveProductFromCart() {
+            document.getElementById("confirm-remove-from-cart").addEventListener("click", function() {
+                if (productId) {
                     fetch('/carrinho/remover', {
                         method: 'DELETE',
                         headers: {
@@ -140,7 +166,7 @@
                         }
                     })
                     .catch(error => console.error('Erro:', error));
-                });
+                }
             });
         }
